requirements.txt
confluent-kafka==2.5.0
pyspark==3.5.1
psycopg2-binary==2.9.9
python-dotenv==1.0.1
ujson==5.10.0